requirements.txt
asgiref==3.8.1
Django==5.2.1
django-unfold==0.59.0
pillow==11.2.1
python-dotenv==1.1.0
shortuuid==1.0.13
sqlparse==0.5.3
